Remove packages mariadb-client, postgresql-client and sqlite.

+++ Dockerfile
@@ -31,9 +31,8 @@ ENV UWSGI_PLUGIN_NAME=python3 WWW_VARDIR=/var/www
 # Update Debian package repository index and install binary packages
 ENV DEBIAN_FRONTEND=noninteractive
 RUN apt-get update -qy && apt-get install --no-install-recommends -qy \
-build-essential libpq-dev mariadb-client postgresql-client python3 \
-python3-dev python3-pip sqlite3 uwsgi uwsgi-plugin-python3 && \
-rm -rf /var/lib/apt/lists/*
+build-essential libpq-dev python3 python3-dev python3-pip uwsgi \
+uwsgi-plugin-python3 && rm -rf /var/lib/apt/lists/*
 
 # Install PyPI packages
 RUN pip3 install pipenv
